simplification of Bragi
The main goal here is to make Bragi more maintainable, mostly through a
more straightforward use of Warp that does not require to maintain several
macros per endpoint.

I've also changed the `SearchDocuments` trait so that it doesn't enforce
a return type for the JSON result. This resolves several cases where we
had to convert to a `Value` and then convert to another type (which
neither performant nor optimal).