Add search endpoint and alt names (#8)
* simplification of Bragi

The main goal here is to make Bragi more maintainable, mostly through a
more straightforward use of Warp that does not require to maintain several
macros per endpoint.

I've also changed the `SearchDocuments` trait so that it doesn't enforce
a return type for the JSON result. This resolves several cases where we
had to convert to a `Value` and then convert to another type (which
neither performant nor optimal).

* bragi: simplify routes.rs by removing redundant filters

* simplification of Bragi

The main goal here is to make Bragi more maintainable, mostly through a
more straightforward use of Warp that does not require to maintain several
macros per endpoint.

I've also changed the `SearchDocuments` trait so that it doesn't enforce
a return type for the JSON result. This resolves several cases where we
had to convert to a `Value` and then convert to another type (which
neither performant nor optimal).

* add a `search` parameter to /autocomplete endpoint

* search: use "phrase" query type

* replace search parameter with /search endpoint

* add alternativve names

* add alt_name and loc_name to query

* update cosmogony to submodules

* add text mapping for country codes

* add alt and local names mapping

* simplification of Bragi

The main goal here is to make Bragi more maintainable, mostly through a
more straightforward use of Warp that does not require to maintain several
macros per endpoint.

I've also changed the `SearchDocuments` trait so that it doesn't enforce
a return type for the JSON result. This resolves several cases where we
had to convert to a `Value` and then convert to another type (which
neither performant nor optimal).

* bragi: simplify routes.rs by removing redundant filters

* fix review comments

* bragi/api.rs: new error message was missing for forward geocoder

* bragi/handlers: fix Context no longer wrapped with an Arc

* workflows/docker: build image in all push events

This should help with forks which can't access SECRETS from a pull
request.

* check docker image builds from PR

* github actions: default value for docker image in pull requests

* update cosmogony version

* update lock file

* fix issue with wrong rebase

* remove wrong type elastic field

* remove disable other mapping and copy alt and loc name to full_label

* allow to retry when getting bad status code from ES during insertion

* add an option to disable refresh

Co-authored-by: sdirollo <sebastien.dirollo@gmail.com>
Co-authored-by: fatal69100 <sebastiendrll@protonmail.com>